+++ randomuser/randomuser/lib/app.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:randomuser/my_home_page.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(primarySwatch: Colors.blue),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
